Switch installer to gradio version
