--- gradio_installer.py
@@ -1,86 +1,0 @@
-import argparse
-import json
-import subprocess
-import sys
-from pathlib import Path
-
-
-def check_version_and_platform() -> bool:
-    version = sys.version_info
-    if version.major != 3 or version.minor < 10:
-        print("ERROR: Python >= 3.10 required")
-        return False
-    if sys.platform not in ("win32", "linux"):
-        print(f"ERROR: unsupported platform {sys.platform}")
-        return False
-    return True
-
-
-def check_git_install() -> bool:
-    try:
-        subprocess.check_call(
-            "git --version",
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            shell=sys.platform == "linux",
-        )
-    except Exception:
-        print("ERROR: git is not installed, please install git")
-        return False
-    return True
-
-
-def create_or_update_venv() -> Path:
-    python = sys.executable
-    subprocess.check_call(f"{python} -m venv venv", shell=sys.platform == "linux")
-    pip_path = Path("venv/Scripts/pip.exe" if sys.platform == "win32" else "venv/bin/pip")
-    subprocess.check_call(f"{pip_path} install -U -r requirements.txt", shell=sys.platform == "linux")
-    return pip_path
-
-
-def install_backend(python: str) -> None:
-    subprocess.check_call(
-        "git submodule update --init --recursive",
-        shell=sys.platform == "linux",
-    )
-    backend_installer = Path("backend/installer.py")
-    if backend_installer.exists():
-        subprocess.check_call(f"{python} {backend_installer} local", shell=sys.platform == "linux")
-
-
-def install() -> None:
-    if not (check_version_and_platform() and check_git_install()):
-        return
-    pip_path = create_or_update_venv()
-    config = Path("config.json")
-    config_dict = json.loads(config.read_text()) if config.exists() else {}
-    install_backend_choice = None
-    while install_backend_choice not in ("y", "n"):
-        install_backend_choice = input("Are you using this locally? (y/n): ").lower()
-    config_dict["run_local"] = install_backend_choice == "y"
-    config.write_text(json.dumps(config_dict, indent=2))
-    if install_backend_choice == "y":
-        python = Path("venv/Scripts/python.exe" if sys.platform == "win32" else "venv/bin/python")
-        install_backend(str(python))
-
-
-def update() -> None:
-    if not check_git_install():
-        return
-    subprocess.check_call("git pull", shell=sys.platform == "linux")
-    subprocess.check_call("git submodule update --init --recursive", shell=sys.platform == "linux")
-    pip_path = Path("venv/Scripts/pip.exe" if sys.platform == "win32" else "venv/bin/pip")
-    if pip_path.exists():
-        subprocess.check_call(f"{pip_path} install -U -r requirements.txt", shell=sys.platform == "linux")
-    else:
-        print("Virtual environment not found. Run install first.")
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Installer for the Gradio LoRA Trainer")
-    parser.add_argument("command", choices=["install", "update"], help="Action to perform")
-    args = parser.parse_args()
-    if args.command == "install":
-        install()
-    else:
-        update()
