+++ install.py
@@ -1,15 +1,13 @@
+import argparse
 import json
-import os
 import subprocess
 import sys
 from pathlib import Path
-
 
 SUPPORTED_PLATFORMS = {"win32", "linux"}
 
 
 def check_environment() -> bool:
-    """Validate python version, platform and git installation."""
     version = sys.version_info
     if version.major != 3 or version.minor < 10:
         print("ERROR: Python >= 3.10 required")
@@ -30,8 +28,7 @@ def check_environment() -> bool:
     return True
 
 
-def create_virtualenv() -> Path:
-    """Create venv and install python requirements."""
+def create_or_update_venv() -> Path:
     python = sys.executable
     subprocess.check_call(f"{python} -m venv venv", shell=sys.platform == "linux")
     pip = Path("venv/Scripts/pip.exe" if sys.platform == "win32" else "venv/bin/pip")
@@ -39,47 +36,46 @@ def create_virtualenv() -> Path:
     return pip
 
 
-def configure_local_run() -> bool:
-    """Ask user if backend should run locally and update config."""
-    config = Path("config.json")
-    data = json.loads(config.read_text()) if config.exists() else {}
+def install_backend(python: str) -> None:
+    subprocess.check_call("git submodule update --init --recursive", shell=sys.platform == "linux")
+    backend_installer = Path("backend/installer.py")
+    if backend_installer.exists():
+        subprocess.check_call(f"{python} {backend_installer} local", shell=sys.platform == "linux")
 
+
+def install() -> None:
+    if not check_environment():
+        return
+    create_or_update_venv()
+    config = Path("config.json")
+    config_dict = json.loads(config.read_text()) if config.exists() else {}
     choice = None
     while choice not in ("y", "n"):
         choice = input("Are you using this locally? (y/n): ").lower()
-
-    run_local = choice == "y"
-    data["run_local"] = run_local
-    config.write_text(json.dumps(data, indent=2))
-    return run_local
-
-
-def update_submodules() -> None:
-    subprocess.check_call("git submodule update --init --recursive", shell=sys.platform == "linux")
+    config_dict["run_local"] = choice == "y"
+    config.write_text(json.dumps(config_dict, indent=2))
+    if choice == "y":
+        python = Path("venv/Scripts/python.exe" if sys.platform == "win32" else "venv/bin/python")
+        install_backend(str(python))
 
 
-def run_backend_installer(python: Path) -> None:
-    """Execute the backend installer inside its directory."""
-    os.chdir("backend")
-    try:
-        subprocess.check_call(f"{python} installer.py local", shell=sys.platform == "linux")
-    finally:
-        os.chdir("..")
-
-
-def main() -> None:
+def update() -> None:
     if not check_environment():
         return
-
-    create_virtualenv()
-    if configure_local_run():
-        update_submodules()
-        python = Path("venv/Scripts/python.exe" if sys.platform == "win32" else "venv/bin/python")
-        if Path("backend/sd_scripts").exists():
-            run_backend_installer(python)
-        else:
-            print("ERROR: backend submodules not found after update.")
+    subprocess.check_call("git pull", shell=sys.platform == "linux")
+    subprocess.check_call("git submodule update --init --recursive", shell=sys.platform == "linux")
+    pip = Path("venv/Scripts/pip.exe" if sys.platform == "win32" else "venv/bin/pip")
+    if pip.exists():
+        subprocess.check_call(f"{pip} install -U -r requirements.txt", shell=sys.platform == "linux")
+    else:
+        print("Virtual environment not found. Run install first.")
 
 
 if __name__ == "__main__":
-    main()
+    parser = argparse.ArgumentParser(description="Installer for the Gradio LoRA Trainer")
+    parser.add_argument("command", nargs="?", default="install", choices=["install", "update"], help="Action to perform")
+    args = parser.parse_args()
+    if args.command == "install":
+        install()
+    else:
+        update()
