+++ requirements.txt
@@ -1,5 +1,3 @@
-pyside6
-qt-material==2.14
 toml
 requests
 gradio
