+++ run.bat
@@ -3,5 +3,5 @@
 title LoRA Trainer
 cd %~dp0
 call venv\Scripts\activate
-python main.py
+python gradio_app.py
 pause
