+++ run.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 source venv/bin/activate
-python main.py
+python gradio_app.py
